+++ bitcoin/src/consensus/encode.rs
@@ -103,43 +103,6 @@ pub trait Encodable {
     fn consensus_encode<W: io::Write + ?Sized>(&self, writer: &mut W) -> Result<usize, io::Error>;
 }
 
-impl Encodable for String {
-    #[inline]
-    fn consensus_encode<W: io::Write + ?Sized>(&self, w: &mut W) -> Result<usize, io::Error> {
-        let b = self.as_bytes();
-        let vi_len = VarInt(b.len() as u64).consensus_encode(w)?;
-        w.emit_slice(b)?;
-        Ok(vi_len + b.len())
-    }
-}
-
-impl Decodable for String {
-    #[inline]
-    fn consensus_decode<R: io::Read + ?Sized>(r: &mut R) -> Result<String, Error> {
-        String::from_utf8(Decodable::consensus_decode(r)?)
-            .map_err(|_| self::Error::ParseFailed("String was not valid UTF8"))
-    }
-}
-
-impl Encodable for Cow<'static, str> {
-    #[inline]
-    fn consensus_encode<W: io::Write + ?Sized>(&self, w: &mut W) -> Result<usize, io::Error> {
-        let b = self.as_bytes();
-        let vi_len = VarInt(b.len() as u64).consensus_encode(w)?;
-        w.emit_slice(b)?;
-        Ok(vi_len + b.len())
-    }
-}
-
-impl Decodable for Cow<'static, str> {
-    #[inline]
-    fn consensus_decode<R: io::Read + ?Sized>(r: &mut R) -> Result<Cow<'static, str>, Error> {
-        String::from_utf8(Decodable::consensus_decode(r)?)
-            .map_err(|_| self::Error::ParseFailed("String was not valid UTF8"))
-            .map(Cow::Owned)
-    }
-}
-
 macro_rules! impl_array {
     ( $size:literal ) => {
         impl Encodable for [u8; $size] {
